ACP-L3-DEF-CONSOLIDATE-CATALOG-PATHS: fix hidden construction and package-shape review findings
Eliminate the second catalog.Service construction inside NewCatalogPathsService
by calling catalog-owned named-factory resolution directly (the exact helper
catalog.Service.ResolveNamedFactory itself delegates to) and by consuming the
already-composed CurrentFactoryDirectoryResolver Wire value instead of
re-deriving it.

Move CatalogPathsService's interface declaration off the crowded
factory_definitions root (which pkgstructurecheck's deletion-only baseline
forbids growing) into the factory_definitions/wire package, which the
service-root-interface-count rule never scans. Chat Sessions now depends on
its own locally-declared 2-method interface instead of a Factory-Definitions-
published bundling type. Fold the new implementation/test files into existing
files in the same packages so pkgfilecountcheck's deletion-only baseline also
shows zero growth (confirmed via A/B diff against origin/main: both gates'
output is now byte-for-byte identical to main).
