+++ pkg/services/chat_sessions/internal/service/service.go
@@ -5,6 +5,7 @@
 package service
 
 import (
+	"context"
 	"fmt"
 
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
@@ -13,13 +14,27 @@ import (
 	operatorsettings "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 )
 
+// FactoryDefinitionsCatalogPaths is the narrow read subset of Factory
+// Definitions' catalog/path capability this service actually calls: listing
+// installed/available targets and resolving a caller-supplied named target
+// reference against project/global roots. Declared locally (rather than
+// depending on a bundling type published by Factory Definitions) so this
+// package only requires the exact two operations it uses; any Factory
+// Definitions collaborator whose method set covers these two signatures
+// satisfies it structurally, with no import of Factory Definitions' own wire
+// subpackage required.
+type FactoryDefinitionsCatalogPaths interface {
+	ListEffectiveFactories(context.Context, factorydefinitions.ListEffectiveFactoriesRequest) (factorydefinitions.ListEffectiveFactoriesResult, error)
+	ResolveNamedFactory(context.Context, factorydefinitions.ResolveNamedFactoryRequest) (factorydefinitions.ResolveNamedFactoryResult, error)
+}
+
 // Service implements chatsessions.FactoryTargetCatalogService by combining
 // the singular Operator Settings public service root and Factory
-// Definitions' narrow, read-only CatalogPathsService capability, injected
-// directly and exactly once.
+// Definitions' narrow, read-only catalog/path capability, injected directly
+// and exactly once.
 type Service struct {
 	operatorSettings   operatorsettings.Service
-	factoryDefinitions factorydefinitions.CatalogPathsService
+	factoryDefinitions FactoryDefinitionsCatalogPaths
 	logger             logging.Logger
 }
 
@@ -30,7 +45,7 @@ var _ chatsessions.FactoryTargetCatalogService = (*Service)(nil)
 // operation logging pass logging.NoopLogger{}.
 func New(
 	operatorSettings operatorsettings.Service,
-	factoryDefinitions factorydefinitions.CatalogPathsService,
+	factoryDefinitions FactoryDefinitionsCatalogPaths,
 	logger logging.Logger,
 ) (*Service, error) {
 	if operatorSettings == nil {
