+++ pkg/services/factory_definitions/wire/catalog_ports.go
@@ -1,10 +1,11 @@
 package wire
 
 import (
+	"context"
+
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factorydefinitionsinternal "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal"
-	catalog "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/catalog"
 	catalogwire "github.com/portpowered/infinite-you/pkg/services/factory_definitions/internal/services/catalog/wire"
 )
 
@@ -65,33 +66,67 @@ func ResolveCurrent(
 	return catalogwire.ResolveCurrent(paths, rootDir)
 }
 
+// CatalogPathsService is Factory Definitions' narrow, stateless, read-only
+// capability for available Factory target metadata and named/current Factory
+// location resolution. It intentionally excludes authoring, compilation,
+// validation, snapshot, distribution, runtime, and session operations; peers
+// that need only catalog and path reads depend on this interface instead of
+// the full factorydefinitions.Service root.
+//
+// This capability is published here, at the Wire composition boundary,
+// rather than at the crowded factory_definitions root: the root package
+// already carries pre-existing, deletion-only service-root-interface-count
+// debt (docs/internal/baselines/package-structure-baseline.json), and that
+// baseline is a strict ratchet -- new entries and count increases are
+// rejected outright, so a new root-level bundling interface is not an option
+// here. Peers depend on this type through their own locally declared narrow
+// interface (see pkg/services/chat_sessions/internal/service's
+// FactoryDefinitionsCatalogPaths), which this concrete implementation
+// satisfies structurally without either side needing to import the other's
+// wire subpackage.
+type CatalogPathsService interface {
+	ListEffectiveFactories(context.Context, factorydefinitions.ListEffectiveFactoriesRequest) (factorydefinitions.ListEffectiveFactoriesResult, error)
+	ResolveNamedFactory(context.Context, factorydefinitions.ResolveNamedFactoryRequest) (factorydefinitions.ResolveNamedFactoryResult, error)
+	ResolveCurrentFactoryLocation(context.Context, factorydefinitions.ResolveCurrentFactoryLocationRequest) (factorydefinitions.ResolveCurrentFactoryLocationResult, error)
+}
+
 // NewCatalogPathsService constructs the narrow, read-only Factory
-// Definitions catalog/path capability from the exact effective-catalog
-// operation and catalog-owned path/filesystem ports used by Wire
-// composition. It reuses the same private catalog collaborator the root
-// Service's ResolveNamedFactory delegates to, so results are identical, and
-// performs no filesystem reads or writes at construction time. logger is the
-// direct, required operation-logging abstraction; callers with no operation
-// logging pass logging.NoopLogger{}.
+// Definitions catalog/path capability from the exact already-composed
+// effective-catalog, named-path, and current-directory collaborators used by
+// canonical Wire composition. Named-Factory resolution calls the catalog
+// subservice's own named-factory helper directly (the same one
+// catalog.Service.ResolveNamedFactory delegates to) instead of constructing a
+// second catalog.Service instance, and current-directory resolution reuses
+// the exact factorydefinitions.CurrentFactoryDirectoryResolver value
+// canonical Wire already constructs once for the rest of the graph.
+// Construction performs no filesystem reads or writes. logger is the direct,
+// required operation-logging abstraction; callers with no operation logging
+// pass logging.NoopLogger{}.
 func NewCatalogPathsService(
 	listEffective factorydefinitions.EffectiveFactoryCatalogOperation,
 	namedPaths factorydefinitions.NamedPathResolver,
-	namedFactoryCatalogFileSystem factorydefinitions.NamedFactoryCatalogFileSystem,
+	resolveCurrentDir factorydefinitions.CurrentFactoryDirectoryResolver,
 	logger logging.Logger,
-) (factorydefinitions.CatalogPathsService, error) {
-	catalogService, err := catalogwire.NewService(catalog.Dependencies{
-		Paths:      namedPaths,
-		FileSystem: namedFactoryCatalogFileSystem,
-	})
-	if err != nil {
-		return nil, err
-	}
-	resolveCurrentDir := func(rootDir string) (string, error) {
-		return catalogwire.ResolveCurrent(namedPaths, rootDir)
+) (CatalogPathsService, error) {
+	resolveNamedFactory := func(
+		ctx context.Context,
+		request factorydefinitions.ResolveNamedFactoryRequest,
+	) (factorydefinitions.ResolveNamedFactoryResult, error) {
+		if err := ctx.Err(); err != nil {
+			return factorydefinitions.ResolveNamedFactoryResult{}, err
+		}
+		resolution, err := catalogwire.ResolveNamedFactory(namedPaths, request.ProjectRoot, request.GlobalRoot, request.Name)
+		if err != nil {
+			return factorydefinitions.ResolveNamedFactoryResult{}, err
+		}
+		if resolution == nil {
+			return factorydefinitions.ResolveNamedFactoryResult{}, factorydefinitions.ErrNamedFactoryNotFound
+		}
+		return factorydefinitions.ResolveNamedFactoryResult{Resolution: *resolution}, nil
 	}
 	return factorydefinitionsinternal.NewCatalogPathsService(
 		listEffective,
-		catalogService.ResolveNamedFactory,
+		resolveNamedFactory,
 		resolveCurrentDir,
 		logger,
 	)
