+++ pkg/wire/chat_sessions.go
@@ -4,7 +4,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
 	chatsessions "github.com/portpowered/infinite-you/pkg/services/chat_sessions"
 	chatsessionswire "github.com/portpowered/infinite-you/pkg/services/chat_sessions/wire"
-	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	factorydefinitionswire "github.com/portpowered/infinite-you/pkg/services/factory_definitions/wire"
 	operatorsettings "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 )
 
@@ -12,12 +12,12 @@ import (
 // Sessions Factory target-catalog root through the owning Chat Sessions Wire
 // package's focused constructor, from the same singular Operator Settings
 // public service root and Factory Definitions' narrow, read-only
-// CatalogPathsService capability the rest of this graph composes. It
-// performs no dependency-bag or service-locator composition: every
-// collaborator is a direct constructor parameter.
+// catalog/path capability the rest of this graph composes. It performs no
+// dependency-bag or service-locator composition: every collaborator is a
+// direct constructor parameter.
 func provideChatSessionsFactoryTargetCatalogService(
 	operatorSettings operatorsettings.Service,
-	catalogPaths factorydefinitions.CatalogPathsService,
+	catalogPaths factorydefinitionswire.CatalogPathsService,
 	logger logging.Logger,
 ) (chatsessions.FactoryTargetCatalogService, error) {
 	return chatsessionswire.NewFactoryTargetCatalogService(operatorSettings, catalogPaths, logger)
