+++ pkg/wire/wire_gen.go
@@ -493,7 +493,7 @@ func InjectBundle(ctx context.Context, edges2 edges.Edges) (*application.Process
 	if err != nil {
 		return nil, err
 	}
-	catalogPathsService, err := wire3.NewCatalogPathsService(effectiveFactoryCatalogOperation, namedPathResolver, namedFactoryCatalogFileSystem, loggingLogger)
+	catalogPathsService, err := wire3.NewCatalogPathsService(effectiveFactoryCatalogOperation, namedPathResolver, v4, loggingLogger)
 	if err != nil {
 		return nil, err
 	}
